requirements-extra.txt
chumpy
gsplat @ git+https://github.com/nerfstudio-project/gsplat.git@v1.3.0
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast